Update ImageConverter logic and add Star Wars retrieval
Modified the `ImageConverter` class to check if `input` starts
with `Banners` instead of checking if it contains `Banners`
before appending the prefix. This ensures the prefix is only
added when necessary.

In the `Program` class, added a call to retrieve information
about the "Star Wars" series and store it in the `starwars`
variable. Also added a clarifying comment for the existing
"Game of Thrones" retrieval code.